Use new header styles in Discover
