+++ src/sentry/static/sentry/app/components/organizations/header.jsx
@@ -1,0 +1,11 @@
+import styled from 'react-emotion';
+import {Flex} from 'grid-emotion';
+
+const Header = styled(Flex)`
+  border-bottom: 1px solid ${p => p.theme.borderDark};
+  box-shadow: ${p => p.theme.dropShadowLight};
+  font-size: 18px;
+  height: 60px;
+`;
+
+export default Header;
